- added comentlikes and uploading Images

+++ MMFoodIdea/Controllers/RecipesController.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using MMFI_Data.Data;
 using MMFI_Entites.Models;
 using MMFI_Entites.ViewModels;
@@ -41,7 +42,12 @@ namespace MMFoodIdea.Controllers
                 cvm.UserId = currentUser.Id;
             }
 
-            cvm.comments = _cServices.GetAllComments(id);
+            cvm.comments = _cServices.GetAllComments(id);            
+
+            foreach(var c in cvm.comments)
+            {
+                c.Likes = await _appDb.CommentLikes.Where(x => x.CommentId == c.CommentID).ToListAsync();
+            }
 
             return View(cvm);
         }
@@ -77,15 +83,48 @@ namespace MMFoodIdea.Controllers
 
         [Authorize]
         public async Task<IActionResult> DeleteComment(Comment comment)
+        {           
+                if (ModelState.IsValid)
+                {
+                    if (comment.UserId == _userManager.GetUserId(User))
+                        await _cServices.DeleteComment(comment);
+                   
+                    return Ok();
+                }
+            
+
+            return View("Error");
+        }
+
+        public async Task<IActionResult> EditComment(Comment comment)
         {
             if (ModelState.IsValid)
             {
-                await _cServices.DeleteComment(comment);
+                _appDb.Comments.Remove(comment);
+                await _appDb.Comments.AddAsync(comment);
+                await _appDb.SaveChangesAsync();
                 return Ok();
             }
 
             return View("Error");
         }
+
+        [Authorize]
+        public async Task<IActionResult> OnLikeClick(Comment comment)
+        {
+            if (ModelState.IsValid)
+            {
+                var userId = _userManager.GetUserId(User);
+
+                await _cServices.CommentLiking(comment, userId);
+                
+                
+                return Ok();
+            }
+
+            return View("Error");
+        }
+
 
 
     }
